- Move several public utilities into internal utilities to consolidate - Delete utilities that weren't being used (dead code) - Deprecated and redirect internally several public utilities that shouldn't have been public (maintain backwards compatibility, nobody's ever really gone in Java) - Add citation, license for modified Apache code
